chore: split function into files and fix norm
