+++ Makefile
@@ -6,7 +6,7 @@
 #    By: pmarkaid <pmarkaid@student.hive.fi>        +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2024/08/01 15:11:43 by pmarkaid          #+#    #+#              #
-#    Updated: 2024/08/10 13:08:09 by pmarkaid         ###   ########.fr        #
+#    Updated: 2024/08/13 23:28:18 by pmarkaid         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -27,9 +27,11 @@ SRCS_FILES = \
 	signals.c 	\
 	bu_func.c 	\
 	validation.c \
+	validation_utils.c \
 	execution.c \
 	execution_utils.c \
-	error.c
+	error.c \
+	dup.c
 
 SRC_DIR = src/
 SRCS = $(addprefix $(SRC_DIR), $(SRCS_FILES))
